+++ Runtime/sources/segment_tree.c
@@ -3,44 +3,44 @@
 
 #include "segment_tree.h"
 
-static void mark(uint32_t curr, uint32_t currBegin, uint32_t currEnd, uint32_t needBegin, uint32_t needEnd);
-static uint32_t sum(uint32_t curr, uint32_t currBegin, uint32_t currEnd, uint32_t needBegin, uint32_t needEnd);
+static void mark(uint64_t curr, uint64_t currBegin, uint64_t currEnd, uint64_t needBegin, uint64_t needEnd);
+static uint64_t sum(uint64_t curr, uint64_t currBegin, uint64_t currEnd, uint64_t needBegin, uint64_t needEnd);
 
-void buildSegmentTree(uint32_t n)
+void buildSegmentTree(uint64_t n)
 {
 	memMngr.segmentTree->currN = n;
-	uint32_t treeSize = 4 * memMngr.segmentTree->currN / memMngr.segmentLen;
-	uint32_t i = 0;
+    uint64_t treeSize = 4 * memMngr.segmentTree->currN / memMngr.segmentLen;
+    uint64_t i = 0;
 
 	//Кажется, можно просто обнулять. Но пока работает)))
 	for (i = 0; i < treeSize; ++i)
 		memMngr.segmentTree->tree[i] = -1;
 
-	memset(memMngr.segmentTree->elements, 0, sizeof(uint32_t) * memMngr.segmentTree->currN);
+    memset(memMngr.segmentTree->elements, 0, sizeof(uint64_t) * memMngr.segmentTree->currN);
 }
 
 void clearSegmentTree()
 {
-	uint32_t treeSize = 4 * memMngr.segmentTree->currN / memMngr.segmentLen;
-	uint32_t i = 0;
+    uint64_t treeSize = 4 * memMngr.segmentTree->currN / memMngr.segmentLen;
+    uint64_t i = 0;
 
 	for (i = 0; i < treeSize; ++i)
 		memMngr.segmentTree->tree[i] = -1;
 
-	memset(memMngr.segmentTree->elements, 0, sizeof(uint32_t) * memMngr.segmentTree->currN);
+    memset(memMngr.segmentTree->elements, 0, sizeof(uint64_t) * memMngr.segmentTree->currN);
 }
 
-void markInSegmentTree(uint32_t begin, uint32_t end)
+void markInSegmentTree(uint64_t begin, uint64_t end)
 {
 	mark(1, 0, memMngr.segmentTree->currN - 1, begin, end);
 }
 
-uint32_t sumInSegmentTree(uint32_t begin, uint32_t end)
+uint64_t sumInSegmentTree(uint64_t begin, uint64_t end)
 {
 	sum(1, 0, memMngr.segmentTree->currN - 1, begin, end);
 }
 
-static uint32_t max(uint32_t a, uint32_t b)
+static uint64_t max(uint64_t a, uint64_t b)
 {
 	if (a > b)
 		return a;
@@ -48,7 +48,7 @@ static uint32_t max(uint32_t a, uint32_t b)
 	return b;
 }
 
-static uint32_t min(uint32_t a, uint32_t b)
+static uint64_t min(uint64_t a, uint64_t b)
 {
 	if (a > b)
 		return b;
@@ -56,10 +56,10 @@ static uint32_t min(uint32_t a, uint32_t b)
 	return a;
 }
 
-static void markElements(uint32_t needBegin, uint32_t needEnd, int32_t mark)
+static void markElements(uint64_t needBegin, uint64_t needEnd, int32_t mark)
 {
 	int i = 0;
-	uint32_t segmentNum = needBegin / memMngr.segmentLen;
+    uint64_t segmentNum = needBegin / memMngr.segmentLen;
 
 	for (i = segmentNum * memMngr.segmentLen; i < (segmentNum + 1) * memMngr.segmentLen; ++i)
 	{
@@ -70,10 +70,10 @@ static void markElements(uint32_t needBegin, uint32_t needEnd, int32_t mark)
 	}
 }
 
-static uint32_t sumElements(uint32_t needBegin, uint32_t needEnd)
+static uint64_t sumElements(uint64_t needBegin, uint64_t needEnd)
 {
-	uint32_t i = 0;
-	uint32_t sum = 0;
+    uint64_t i = 0;
+    uint64_t sum = 0;
 
 	for (i = needBegin; i <= needEnd; ++i)
 		sum += memMngr.segmentTree->elements[i];
@@ -81,7 +81,7 @@ static uint32_t sumElements(uint32_t needBegin, uint32_t needEnd)
 	return sum;
 }
 
-static void push(uint32_t curr)
+static void push(uint64_t curr)
 {
 	if (memMngr.segmentTree->tree[curr] == -1)
 				return;
@@ -91,17 +91,17 @@ static void push(uint32_t curr)
 	memMngr.segmentTree->tree[curr] = -1;
 }
 
-static uint32_t left(uint32_t val)
+static uint64_t left(uint64_t val)
 {
 	return val * memMngr.segmentLen;
 }
 
-static uint32_t right(uint32_t val)
+static uint64_t right(uint64_t val)
 {
 	return (val + 1) * memMngr.segmentLen - 1;
 }
 
-static void mark(uint32_t curr, uint32_t currBegin, uint32_t currEnd, uint32_t needBegin, uint32_t needEnd)
+static void mark(uint64_t curr, uint64_t currBegin, uint64_t currEnd, uint64_t needBegin, uint64_t needEnd)
 {
 	if (needBegin > needEnd)
 		return;
@@ -116,7 +116,7 @@ static void mark(uint32_t curr, uint32_t currBegin, uint32_t currEnd, uint32_t n
 	//Дошли до листа
 	if (currBegin == currEnd)
 	{
-		uint32_t label = 0;
+        uint64_t label = 0;
 
 		if (memMngr.segmentTree->tree[currBegin] > 0)
 			label = 1;
@@ -128,13 +128,13 @@ static void mark(uint32_t curr, uint32_t currBegin, uint32_t currEnd, uint32_t n
 
 	push(curr);
 
-	uint32_t currMiddle = (currBegin + currEnd) / 2;
+    uint64_t currMiddle = (currBegin + currEnd) / 2;
 
 	mark(curr * 2, currBegin, currMiddle, needBegin, min(right(currMiddle), needEnd));
 	mark(curr * 2 + 1, currMiddle + 1, currEnd, max(left(currMiddle + 1), needBegin), needEnd);
 }
 
-static uint32_t sum(uint32_t curr, uint32_t currBegin, uint32_t currEnd, uint32_t needBegin, uint32_t needEnd)
+static uint64_t sum(uint64_t curr, uint64_t currBegin, uint64_t currEnd, uint64_t needBegin, uint64_t needEnd)
 {
 	if (needBegin > needEnd)
 		return 0;
@@ -146,7 +146,7 @@ static uint32_t sum(uint32_t curr, uint32_t currBegin, uint32_t currEnd, uint32_
 	if (currBegin == currEnd)
 		return sumElements(needBegin, needEnd);
 
-	uint32_t currMiddle = (currBegin + currEnd) / 2;
+    uint64_t currMiddle = (currBegin + currEnd) / 2;
 
 	return sum(curr * 2, currBegin, currMiddle, needBegin, min(right(currMiddle), needEnd)) +
 		sum(curr * 2 + 1, currMiddle + 1, currEnd, max(left(currMiddle + 1), needBegin), needEnd);
